fe s: rename: loginPassword > login, endpoint utils
