+++ shared/src/endpoints/enroll.ts
@@ -1,5 +1,5 @@
 import * as yup from "yup";
-import { thenReq, thenUrlReq } from "./utils";
+import { thenReq, thenUrlReq } from "./__utils";
 
 /**
  * Contacts
