+++ shared/src/endpoints/index.ts
@@ -1,6 +1,7 @@
 export * from "./enroll";
 export * from "./is-user-enrolled";
 export * from "./login-email";
+export * from "./login";
 export * from "./pay-confirm";
 export * from "./pay-get-payment-info";
 export * from "./pay";
