+++ shared/src/endpoints/login-email.ts
@@ -1,5 +1,5 @@
 import * as yup from "yup";
-import { emailSchema } from "./utils";
+import { emailSchema } from "./__utils";
 
 /**
  * LoginEmail
