+++ shared/src/endpoints/pay.ts
@@ -1,5 +1,5 @@
 import * as yup from "yup";
-import { thenReq, thenNull, cfSchema } from "./utils";
+import { thenReq, thenNull, cfSchema } from "./__utils";
 import { BillingOptions } from "../types";
 
 /**
