+++ shared/src/endpoints/user-exists.ts
@@ -1,5 +1,5 @@
 import * as yup from "yup";
-import { emailSchema } from "./utils";
+import { emailSchema } from "./__utils";
 
 //
